+++ Tests.Live.DofusDb.ApiClients/JobImagesClientTest.cs
@@ -7,7 +7,7 @@ public class JobImagesClientTest
     [Fact(Skip = "The job images API doesn't seem to work.")]
     public async Task JobImagesClient_Should_GetItemImage()
     {
-        IDofusDbImageClient<int> client = DofusDbClient.Beta(Constants.Referrer).JobImages();
+        IDofusDbImageClient<long> client = DofusDbClient.Beta(Constants.Referrer).JobImages();
         await using Stream imageStream = await client.GetImageAsync(11);
         await Verify(imageStream, "jpg");
     }
